rename main_menu to main
