+++ src/main.cpp
@@ -1,7 +1,7 @@
 #include "modules/base/base.h"
 #include "modules/engine/engine.h"
 #include "modules/lua/lua.h"
-#include "modules/main_menu/main_menu.h"
+#include "modules/main/main_menu.h"
 #include "modules/rocket/rocket_launch.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
@@ -20,15 +20,15 @@ int main(void) {
 
   load_config_file();
 
-  world.import<BaseModule>();
-  world.import<LuaModule>();
-  world.import<EngineModule>();
-  world.import<StatsModule>();
-  world.import<SimulationModule>();
-  world.import<MainMenuModule>();
-  world.import<SiteModule>();
-  world.import<RocketLaunchModule>();
-  world.import<StaffModule>();
+  world.import <BaseModule>();
+  world.import <LuaModule>();
+  world.import <EngineModule>();
+  world.import <StatsModule>();
+  world.import <SimulationModule>();
+  world.import <MainMenuModule>();
+  world.import <SiteModule>();
+  world.import <RocketLaunchModule>();
+  world.import <StaffModule>();
 
   // Main Loop
   logger->info("Starting");
